+++ src/base/b-bottom-slide/test/helpers.js
@@ -12,10 +12,13 @@ const
 	h = include('tests/helpers');
 
 /**
+ * Initializes a bottom slide component to test
+ *
  * @param {object} props
  * @param {object} content
+ * @returns {!Promise<[!Object, !Element]>}
  */
-async function init(props = {}, content = undefined) {
+async function initBottomSlide(props = {}, content = undefined) {
 	content = {
 		default: content ?? {
 			tag: 'div',
@@ -52,63 +55,78 @@ async function init(props = {}, content = undefined) {
 	return [component, componentNode];
 }
 
-module.exports.init = init;
+module.exports.initBottomSlide = initBottomSlide;
 
 /**
- * @param {object} c
+ * Invokes `open` of the specified component
+ *
+ * @param {!Object} component
  * @param {number} [step]
+ * @returns {!Promise<void>}
  */
-async function open(c, step = undefined) {
+async function open(component, step = undefined) {
 	const
 		{page} = globalThis._testEnv;
 
-	await c.evaluate((ctx, [step]) => ctx.open(step), [step]);
+	await component.evaluate((ctx, [step]) => ctx.open(step), [step]);
 	await h.bom.waitForIdleCallback(page);
 }
 
 module.exports.open = open;
 
 /**
- * @param {object} c
+ * Invokes `close` of the specified component
+ *
+ * @param {Object} component
+ * @returns {!Promise<void>}
  */
- async function close(c) {
+ async function close(component) {
 	const
 		{page} = globalThis._testEnv;
 
-	await c.evaluate((ctx) => ctx.close());
+	await component.evaluate((ctx) => ctx.close());
 	await h.bom.waitForIdleCallback(page);
 }
 
 module.exports.close = close;
 
 /**
- * @param {object} c
+ * Invokes `next` of the specified component
+ *
+ * @param {!Object} component
+ * @returns {!Promise<void>}
  */
- async function next(c) {
+ async function next(component) {
 	const
 		{page} = globalThis._testEnv;
 
-	await c.evaluate((ctx) => ctx.next());
+	await component.evaluate((ctx) => ctx.next());
 	await h.bom.waitForIdleCallback(page);
 }
 
 module.exports.next = next;
 
 /**
- * @param {object} c
+ * Invokes `prev` of the specified component
+ *
+ * @param {!Object} component
+ * @returns {!Promise<void>}
  */
- async function prev(c) {
+async function prev(component) {
 	const
 		{page} = globalThis._testEnv;
 
-	await c.evaluate((ctx) => ctx.prev());
+	await component.evaluate((ctx) => ctx.prev());
 	await h.bom.waitForIdleCallback(page);
 }
 
 module.exports.prev = prev;
 
 /**
- * @param {number} percent
+ * Returns a value of the global window height
+ *
+ * @param {number} [percent] - percent of the resulting height to return
+ * @returns {!Promise<number>}
  */
 function getAbsolutePageHeight(percent = 100) {
 	const
@@ -120,28 +138,37 @@ function getAbsolutePageHeight(percent = 100) {
 module.exports.getAbsolutePageHeight = getAbsolutePageHeight;
 
 /**
- * @param {object} c
+ * Returns an offset between the global window and component window element
+ *
+ * @param {Object} component
+ * @returns {!Promise<number>}
  */
-function getAbsoluteComponentWindowOffset(c) {
-	return c.evaluate((ctx) => Math.round(globalThis.innerHeight - ctx.block.element('window').getBoundingClientRect().y));
+function getAbsoluteComponentWindowOffset(component) {
+	return component.evaluate((ctx) => Math.round(globalThis.innerHeight - ctx.block.element('window').getBoundingClientRect().y));
 }
 
 module.exports.getAbsoluteComponentWindowOffset = getAbsoluteComponentWindowOffset;
 
 /**
- * @param {object} c
+ * Returns `offsetHeight` of the `window` element from the specified component
+ *
+ * @param {!Object} component
+ * @returns {!Promise<number>}
  */
-function getAbsoluteComponentWindowHeight(c) {
-	return c.evaluate((ctx) => ctx.block.element('window').offsetHeight);
+function getAbsoluteComponentWindowHeight(component) {
+	return component.evaluate((ctx) => ctx.block.element('window').offsetHeight);
 }
 
 module.exports.getAbsoluteComponentWindowHeight = getAbsoluteComponentWindowHeight;
 
 /**
- * @param {object} c
+ * Returns a `y` position from the component `boundedClientRect`
+ *
+ * @param {!Object} component
+ * @returns {!Promise<number>}
  */
-function getComponentWindowYPos(c) {
-	return c.evaluate((ctx) => Math.round(ctx.block.element('window').getBoundingClientRect().y));
+function getComponentWindowYPos(component) {
+	return component.evaluate((ctx) => Math.round(ctx.block.element('window').getBoundingClientRect().y));
 }
 
 module.exports.getComponentWindowYPos = getComponentWindowYPos;